Wrapping up the onboarding process. There's a lot of changes in this set, too many for me to really explain in this bleeding commit message. Or remember.

+++ DolomiteWcfService/LocalStorageManager.cs
@@ -88,7 +88,11 @@ namespace DolomiteWcfService
         public void DeleteFile(string filename)
         {
             // Delete the file
-            File.Delete(GetPath(filename));
+            string filePath = GetPath(filename);
+            if (File.Exists(filePath))
+            {
+                File.Delete(GetPath(filename));
+            }
         }
 
         #endregion
